Move cross window communication aspects into a separate file
 - Similar to last commit, separating logic for these features so that they're easier to maintain and reason about
